+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file. This projec
 
 ### Added
 
-- Library overview: Added average track length and average play count
+- Library overview: Added average track length, average play count, average file size, and total library size
 
 ## 1.0.0 - 2025-07-05